feat: add unique ID tracking to motioncontroller server
ZMQ connection type is changed from REP/REQ to PAIR.
The syntax is also updated to be more SCPI-like.
Each motion added to the queue from the motor plugin is now given a unique identifier generated with the uuid package, which is used to detect motion completion unambiguously.
All client-side code for communication with the server is moved to a new class, ManiClient, so that the actual motor code is much more readable.